+++ ulysse/server.py
@@ -4,62 +4,31 @@ from sanic import Sanic
 from sanic.response import json, text
 from sanic.exceptions import ServerError
 
-from .constants import DEFAULT_COLUMNS
-from .database import db
-from .utils import _generate_score_key
+from .database import decode_siret, retrieve_siret, retrieve_sirets
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
 
-def _get_sirens(column_name, column_value):
-    """Return a list of `siren` codes corresponding to the match."""
-    score_key = _generate_score_key(column_name, column_value)
-    score = db.get(score_key)
-    if not score:
-        msg = '😢 No entry found for {0}/{1}, is it indexed?'.format(
-            column_name, column_value)
-        logger.error(msg)
-        raise ServerError(msg)
-    return db.zrangebyscore(column_name, score, score)
+def dict_to_csv(row, column_names, separator=';'):
+    """Turn a `row` of dicts into one or many CSV row(s)."""
+    lines = []
+    for datemaj, data in sorted(row.items()):
+        line = separator.join(
+            column_name in data and data[column_name] or ''
+            for column_name in column_names
+        )
+        lines.append(line)
+    return '\n'.join(lines)
 
 
-def _redis_to_dict(siren, column_names, encoding='utf-8'):
-    """Turn a Redis response into a decoded dict."""
-    return {
-        k.decode(encoding): v.decode(encoding)
-        for k, v in db.hgetall(siren).items()
-        if k.decode(encoding) in column_names
-    }
-
-
-def _redis_to_csv(siren, column_names, encoding='utf-8', separator=';'):
-    """Turn a Redis response into a CSV row."""
-    data = db.hgetall(siren)
-    return separator.join([
-        data[str.encode(column_name)].decode(encoding)
-        for column_name in column_names
-    ])
-
-
-def _format_response(sirens, format, column_names):
-    """Given a `format`, fetch, serialize and return the appropriated data."""
+def format_response(rows, format, column_names):
+    """Given a `format`, serialize and return the appropriated data."""
     if format == 'json':
-        try:
-            rows = [_redis_to_dict(siren, column_names) for siren in sirens]
-        except KeyError as e:
-            msg = '⁉️ Retrieving a column that is not stored: {0}'.format(e)
-            logger.error(msg)
-            raise ServerError(msg)
         return json(rows)
     elif format == 'csv':
         headers = ';'.join(column_names) + '\n'
-        try:
-            rows = [_redis_to_csv(siren, column_names) for siren in sirens]
-        except KeyError as e:
-            msg = '⁉️ Retrieving a column that is not stored: {0}'.format(e)
-            logger.error(msg)
-            raise ServerError(msg)
+        rows = [dict_to_csv(row, column_names) for row in rows]
         return text(headers + '\n'.join(rows))
 
 
@@ -68,14 +37,20 @@ app = Sanic(__name__)
 
 @app.route('/<name>/<value>')
 async def server(request, name, value):
+    """Quick and dirty API. Async for fun."""
     limit = int(request.args.get('limit', 3))
+    offset = int(request.args.get('offset', 0))
     format = request.args.get('format', 'json')
-    logger.info('🙇 Requesting {0} items in {1}'.format(limit, format))
+    logger.info('🙇 Requesting {0} (offset={1}) items in {2}'.format(
+                limit, offset, format))
     columns = request.args.get('columns')
-    column_names = columns and columns.split(',') or app.config.COLUMNS
-    sirens = _get_sirens(name, value)
+    column_names = columns and columns.split(',') or app.config.columns
+    sirets = retrieve_sirets(name, value, offset=offset, limit=limit)
+    if not sirets:
+        raise ServerError('😢 No entry found for {0}/{1}'.format(name, value))
     logger.info('🐿 Retrieving {0} results for the {1}/{2} couple'.format(
-                len(sirens), name, value))
-    sirens = sirens[:limit]
-    logger.info('🍫 Returning {0} results in {1}'.format(len(sirens), format))
-    return _format_response(sirens, format, column_names)
+                len(sirets), name, value))
+    logger.info('🍫 Returning {0} results in {1}'.format(len(sirets), format))
+    rows = [decode_siret(retrieve_siret(siret), column_names)
+            for siret in sirets]
+    return format_response(rows, format, column_names)