+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.0.0 — 2016-11-05 — 🗄 History
+
+* Ability to load incremental updates
+* **Breaking-change**: use SIREN + NIC (SIRET) as key for storing data
+* **Breaking-change**: the undocumented low-level API has been refactored
+* **Breaking-change**: returned JSON has now one more level with DATEMAJ
+
+
 ## 1.0.1 — 2016-11-01 — 🐞 Fix parameters parsing
 
 * Better handling of list parameters